+++ src/main/java/com/onenth/OneNth/domain/post/repository/PostRepository.java
@@ -4,8 +4,6 @@ import com.onenth.OneNth.domain.post.entity.Post;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import com.onenth.OneNth.domain.post.entity.enums.PostType;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
