--- src/main/java/com/onenth/OneNth/domain/post/repository/PostScrapRepository.java
@@ -1,9 +1,0 @@
-package com.onenth.OneNth.domain.post.repository;
-
-import com.onenth.OneNth.domain.post.entity.Scrap;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostScrapRepository extends JpaRepository<Scrap, Long> {
-
-    boolean existsByPostIdAndMemberId(Long postId, Long memberId);
-}
