+++ src/main/java/com/onenth/OneNth/domain/post/repository/commentRepository/CommentRepository.java
@@ -1,10 +1,10 @@
-package com.onenth.OneNth.domain.post.repository;
+package com.onenth.OneNth.domain.post.repository.commentRepository;
 
 import com.onenth.OneNth.domain.post.entity.Post;
 import com.onenth.OneNth.domain.post.entity.PostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostCommentRepository  extends JpaRepository<PostComment, Long> {
+public interface CommentRepository extends JpaRepository<PostComment, Long> {
 
     long countByPost(Post post);
